Added user filter middleware
